Ensure deletion of NIC and Disks
Sometimes the disk a n d the NIC don't get deleted. I set the parameter to allow them to be deleted at creation and set it force deletion in thhe Clean-Up function. I also put the VM first, the Vnet, the NSG then finally the RG so it has another chance to delete anything it missed.